Cleanup and document ordering
A few minor code cleanups simplifying the naming and use of `context`.

Reorder the constants at the top and annotate them with my best
understanding of the execution flow. Make the same updates in the
README.

Also add lexing to the README example, and add a note about multiplexing
probably being broken.