Refactor static metric mode
Previously the analyzer could be run in a "standalone" mode (without
instrumentation or tracer) to collect static metrics only (excluding
timings). This refactors that solution by introducing a
`capture_field_timings` option on the instrumentation instead.

This simplifies the analyzer by eliminating the conditional modes; now
it always runs in the same way.